backend/requirements.txt
# TourVisor Bot Dependencies
# Responses API Migration

# HTTP клиент для TourVisor API
httpx>=0.27.0

# Загрузка переменных окружения
python-dotenv>=1.0.0

# OpenAI SDK для Yandex Responses API
openai>=1.50.0

# Web UI
flask>=3.0.0
flask-cors>=4.0.0
